requirements.txt
ijson
geopy
requests
python-dotenv
ratelimit
certifi